% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bplsr-package.R
\docType{package}
\name{bplsr-package}
\alias{bplsr-package}
\title{bplsr: Bayesian partial least squares regression}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Szymon Urbas \email{szymon.urbas@ucd.ie}

}
\keyword{internal}
