% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BPLSR.R
\name{bplsr}
\alias{bplsr}
\title{Run the BPLS regression model}
\usage{
bplsr(
  X,
  Y,
  Xtest = NULL,
  Prior = NULL,
  Qs = NULL,
  N_MCMC = 20000,
  BURN = ceiling(0.3 * N_MCMC),
  Thin = 1,
  model.type = "vanilla",
  scale. = TRUE,
  center. = TRUE,
  ...
)
}
\arguments{
\item{X}{Matrix of predictor variables.}

\item{Y}{Vector or matrix of responses.}

\item{Xtest}{Matrix of predictor variables to predict for.}

\item{Prior}{List of hyperparameters specifying the parameter prior distributions. If left NULL, a generic set of priors will be generated.}

\item{N_MCMC}{Number of iterations to run the Markov chain Monte Carlo algorithm}

\item{BURN}{Number of iteration to be discarded as the burn-in.}

\item{Thin}{Thinning procedure for the MArkov chain. \code{Thin = 1} results in no thinning. Only use for long chains to reduce memory.}

\item{model.type}{Type of BPLS model to use; one of \code{vanilla}, \code{ss} (spike-and-slab), or \text{LASSO}}

\item{scale.}{Logical; if \code{TRUE} then the data variables will be scale to have unit variance.}

\item{...}{Additional arguments for \code{bplsr.predict} if \code{Xtest} is provided.}

\item{center}{Logical; if \code{TRUE} then the data variables will be zero-centred.}
}
\value{
A list of
\item{\code{chain}}{A Markov chain of samples from the parameter posterior.}
\item{\code{X}}{Original set of predictor variables.}
\item{\code{Y}}{Original set of response variables.}
\item{\code{Xtest}}{Original set of predictor variables to predict from; if \code{Xtest} is provided.}
\item{\code{Ytest}}{Point predictions for new responses; if \code{Xtest} is provided.}
\item{\code{Ytest_PI}}{Prediction intervals for new responses (by default 95\\%); if \code{Xtest} is provided.}
\item{\code{Ytest_dist}}{Posterior predictive distributions for new responses; if \code{Xtest} is provided.}
\item{\code{diag}}{Additional diagnostics for assessing chain convergence.}
}
\description{
Run the BPLS regression model
}
