% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{milk_MIR}
\alias{milk_MIR}
\title{Milk traits and corresponding mid-infrared spectra}
\format{
A list of 3 dataframes:
}
\usage{
data(milk_MIR)
}
\description{
Data containing spectral measurements for XX milk samples originating from XX cows along with various chemical and technological traits measures.
}
\examples{
data(milk_MIR, package="bplsr")
}
\references{
Streuli, H. (1973). Der heutige Stand der Kaffee-Chemie, In \emph{Proceedings of the 6th International Colloquium on Coffee Chemistry}, June 4-June 9, 1973, Bogata, Colombia, pp. 61-72. Paris, France: Association Scientifique International du Café (ASIC).
}
\keyword{datasets}
