% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bplsr-package.R
\docType{package}
\name{bplsr-package}
\alias{bplsr-package}
\title{bplsr: Bayesian partial least squares regression}
\description{
Fits the Bayesian partial least squares regression model introduced in Urbas et al. (2024) \doi{10.1214/24-AOAS1947}. Suitable for univariate and multivariate regression with high-dimensional data.
}
\author{
\strong{Maintainer}: Szymon Urbas \email{szymon.urbas@mu.ie}

Other contributors:
\itemize{
  \item Pierre Lovera [contributor]
  \item Robert Daly [contributor]
  \item Alan O'Riordan [contributor]
  \item Donagh Berry [contributor]
  \item Isobel Claire Gormley [contributor]
}

}
\keyword{internal}
