% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BPLSR.R
\name{bplsr.predict}
\alias{bplsr.predict}
\title{Predict from a fitted BPLS regression model}
\usage{
bplsr.predict(model, newdata, PredInterval = 0.95)
}
\arguments{
\item{model}{Output of \code{bplsr}.}

\item{newdata}{Matrix of predictor variables to predict for.}

\item{PredInterval}{Intended coverage of prediction intervals (between 0 and 1). Setting the value to 0 only produces point predictions without prediction intervals.}
}
\value{
A list of:
\item{\code{Ytest}}{Point predictions for new responses; if \code{Xtest} is provided.}
\item{\code{Ytest_PI}}{Prediction intervals for new responses (by default 0.95 coverage); if \code{Xtest} is provided.}
\item{\code{Ytest_dist}}{Posterior predictive distributions for new responses; if \code{Xtest} is provided.}
}
\description{
Generates predictions from the fitted BPLS regression model using Monte Carlo simulation.
}
\details{
Predictions of the responses are generated from the posterior predictive distribution, marginalising out the model parameters; see Section 3.5 of Urbas et Al. (2024+).
}
\references{
Urbas, S., Lovera, P., Daly, R., O'Riordan, A., Berry, D., and Gormley, I. C. (2024). "Predicting milk traits from spectral data using Bayesian probabilistic partial least squares regression." \emph{The Annals of Applied Statistics}, 18(4): 3486-3506. \if{html}{\out{<\doi{10.1214/24-AOAS1947}>}}
}
