% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BPLSR.R
\name{bplsr.predict}
\alias{bplsr.predict}
\title{Predict from a fitted BPLS regression model}
\usage{
bplsr.predict(model, newdata, PredInterval = 0.95)
}
\arguments{
\item{model}{Output of \code{bplsr}.}

\item{newdata}{Matrix of predictor variables to predict for.}

\item{PredInterval}{Intended coverage of prediction intervals (between 0 and 1). Setting the value to 0 only produces point predictions without prediction intervals.}
}
\value{
A list of
\item{\code{Ytest}}{Point predictions for new responses; if \code{Xtest} is provided.}
\item{\code{Ytest_PI}}{Prediction intervals for new responses (by default 95\\%); if \code{Xtest} is provided.}
\item{\code{Ytest_dist}}{Posterior predictive distributions for new responses; if \code{Xtest} is provided.}
}
\description{
Predict from a fitted BPLS regression model
}
